define externalAddress in asset.Wallet
This replaces the separate addressPKH and addressWPKH interface methods
with a single externalAddress method. This matches the single
changeAddress method.

The asset.Wallet implementations, such as rpcWallet and spvWallet, know
their segwit capability and mode. In the case of rpcWallet, it has a
segwit flag so that it can support multiple bitcoind backends with or
without segwit addresses. In the case of spvWallet, it only ever
provides segwit addresses and does all operations assuming segwit,
and it had addressPKH returning an "unimplemented" error.
The case of Electrum will be similar to rpcWallet, with a segwit flag
to indicate what sort of external wallet is being driven.