move the RPC-oriented error helpers to RPC code
This moves the rpcclient-oriented error helpers, isTxNotFoundErr and
isMethodNotFoundErr, into rpcclient.go where they should only be used.
Errors from other wallets such as the native SPV wallet and an Electrum
wallet should not use these functions as they match on the
dcrjson.RPCError type and btcjson error codes, which are only applicable
to a bitcoin core backend and an rpcclient client.

Instead of using isTxNotFoundErr in baseWallet after calling
node.getWalletTransaction to translate to asset.CoinNotFoundError, this
now translates as appropriate for the asset.Wallet implementation.
Namely, in rpcclient.go it uses these existing helper functions, while
in spv.go it checks the WalletTransactionNotFound error kind.

Unfortunately, isMethodNotFoundErr is stil applied in the baseWallet
constructor when calling the estimateFee helper to detect backends
that do not offer the method.  This is not as easily changed because
the estimateFee helper accepts a RawRequester and is inherently
RPC-oriented (but some implementers just ignore that argument).
This is something to refactor another day.