client/asset/btc: add ElectrumExchangeWallet
This creates the ExchangeWalletElectrum type for using an external
Electrum wallet along with an ElectrumX server.

This is intended to support Electrum forks for LTC, BCH, and BTC.

The new walletTypeElectrum is defined as "electrumRPC".

The ExchangeWalletElectrum type overrides some baseWallet methods:
- FindRedemption since the process is very different (and simpler)
  from that used by block filter-enabled wallets.
- feeRate, to avoid pointless conversions and btcjson types
- watchBlocks and reportNewTip, since it is not necessary to retarget a
  starting point for redemption searches.
- Connect, since it sets estimateFee to the overridden feeRate method,
  and starts the overridden watchBlocks instead of the baseWallet
  method.

It also statisfies the Sweeper and FeeRater interfaces.

To power ExchangeWalletElectrum, we define an electrumWallet type
that satisfies the asset.Wallet interface similar to rpcWallet and
spvWallet.

The electrumWallet type is powered by an electrumWalletClient, which
communicates with the external Electrum wallet via RPC, and an
electrumNetworkClient, which communicates with a remote Electrum(X)
server for block headers and verbose transaction data.

The general design of electrumWallet is to govern electrumNetworkClient
so that it tracks the electrum rever in used by the Electrum *wallet*.
This is done for a couple of reasons:
- ensure consistent results between the two
- allow the Electrum wallet to handle server selection

-----

add getBestBlockHeader to asset.Wallet

This introduces to the asset.Wallet interface a getBestBlockHeader
method that can be used in place of getBestBlockHash followed by
blockFromHash.

This method is helpful since not all backends allow querying for
block headers by hash. (I'm looking at you Electrum.)

-----

add a fast checkWalletTx optional wallet backend method

This defines an optional walletTxChecker interface, with a single
checkWalletTx method, to complement the current asset.Wallet interface.
walletTxChecker provide a fast wallet tx query when block info not
needed. This should be treated as an optimization method, where
getWalletTransaction may always be used in its place if it does not
exists.

The difference between getWalletTransaction and checkWalletTx is that
getWalletTransaction will return a more verbose struct containing
details on the block containing the transaction. If it is only
necessary to obtain the raw transaction and a confirmation count,
the checkWalletTx may be used.
This is also motivated by Electrum wallets where it is considerably
more expensive to obtain the verbose result from a network server
than to get the checkWalletTx result from the local wallet DB.

As with getWalletTransaction, checkWalletTx is only intended to work
with wallet related transactions.

-----

remove Details field from GetTransactionResult

This simplifies the GetTransactionResult type in wallettypes.go.
Several fields from the bitcoind/btcwallet response that are both
unneeded and difficult to satisfy are now omitted. e.g. the native
btcwallet and an external Electrum backend would have to jump through
hoops to reproduce certain fields that are not strictly necessary.
Namely, the Amount, Fee, and Details fields. All of the required
data can be obtained by decoding the raw transaction contined in the
Hex field.

tolerate some address reuse where needed