windows build
This sets the window icon, which is different from either the tray icon
or any icon embedded in the binary such as the windows ICOs. On
Linxu, this uses gtk_window_set_icon to set the icon from one of the
embedded (in-memory) images so it does not need the file on disk.
On Windows, it applies the icon in a resource.

This also changes to os-specific systray icons since Windows in
particular needs these to be ico files, not pngs.

This also adds Windows syso resource files for dexc-desktop, as is
already done for dexc with -systray builds. On Windows, these resources
are also used to set the window's icon with LoadImage specifying
the window's handle and resource name instead of a file name.