add new database package
Add the Lexi DB package, which wraps badger DB to provide a simplified
API, to capture the sometimes tedious mechanics that we've repeated
in various places, and to add utilities for indexing data for quick
retrieval of filtered data.

This change is just the addition of the package. I will demonstrate
it's use in tatanka separately, and also show how this can replace
in our client/asset tx DB implementations.